// lib/study-api.ts
export type StudyToken = {
  w: string;
  strongs: string;
  transliteration?: string;
  gloss?: string;
  morph?: string;
};